requirements.txt
click == 5.1
lxml == 3.4.4
requests == 2.8.1
Pillow == 3.0.0
suds_jurko == 0.6
